requirements.txt
SpeechRecognition==3.10.1
pyttsx3==2.90
google-generativeai==0.3.2
PyAudio==0.2.14
webbrowser==0.0.1
